derive Clone on TermInfo struct
This patch adds a clone derivation to the TermInfo struct.
Deep copying is a pretty basic operation, especially in rust
where it often makes sense to simplfy borrow checking issues
by just paying for a few extra cycles, and not having the
ability to clone makes the TermInfo type harder to work with.